+++ fs/dir.go
@@ -4,6 +4,8 @@ import (
 	"errors"
 	"fmt"
 	"io/ioutil"
+	"regexp"
+	"strings"
 	"time"
 )
 
@@ -22,6 +24,44 @@ type DirScanner struct {
 	AllSize      int64
 	TopChildDick map[string]ChildDirData
 	Runtime      time.Duration
+
+	//内部变量
+	//排除名称，以"*"匹配（不包含）
+	excludeExp []string
+	//过滤表达式（包含）
+	includeExp []string
+}
+
+//exclude exp for dir scan
+func (ds *DirScanner) Exclude(excludes ...string) *DirScanner {
+	var newExcludes []string
+	for _, ecld := range excludes {
+		if "" == strings.TrimSpace(ecld) {
+			continue
+		}
+		newExcludes = append(newExcludes, ecld)
+	}
+	if len(newExcludes) == 0 {
+		return ds
+	}
+	ds.excludeExp = append(ds.excludeExp, newExcludes...)
+	return ds
+}
+
+//exclude exp for dir scan
+func (ds *DirScanner) Include(includes ...string) *DirScanner {
+	var newInclude []string
+	for _, icld := range includes {
+		if "" == strings.TrimSpace(icld) {
+			continue
+		}
+		newInclude = append(newInclude, icld)
+	}
+	if len(newInclude) == 0 {
+		return ds
+	}
+	ds.includeExp = append(ds.includeExp, newInclude...)
+	return ds
 }
 
 //to star scan the dir.
@@ -44,6 +84,7 @@ func (ds *DirScanner) Scan() error {
 func (ds *DirScanner) scanRecursion(vDir string) int64 {
 	files, err := ioutil.ReadDir(vDir)
 	if err != nil {
+		fmt.Println(err)
 		return 0
 	}
 	isTopClass := false
@@ -54,6 +95,9 @@ func (ds *DirScanner) scanRecursion(vDir string) int64 {
 	var currentSize int64 = 0
 	for _, fl := range files {
 		name := fl.Name()
+		if ds.ignoreScan(name) {
+			continue
+		}
 		var size int64
 		if fl.IsDir() {
 			ds.AllDirItem += 1
@@ -76,8 +120,58 @@ func (ds *DirScanner) scanRecursion(vDir string) int64 {
 	return currentSize
 }
 
+//ignore scan target name
+func (ds *DirScanner) ignoreScan(name string) bool {
+	ignore := false
+	allExp := "*"
+	if len(ds.includeExp) > 0 {
+		isFilter := false
+		for _, filter := range ds.includeExp {
+			if "" == strings.TrimSpace(filter) {
+				continue
+			}
+			if idx := strings.Index(name, allExp); idx > -1 {
+				filter = strings.ReplaceAll(filter, allExp, ".*")
+				if isMatch, er := regexp.MatchString(filter, name); er == nil && isMatch {
+					isFilter = true
+					break
+				}
+			} else if name == filter {
+				isFilter = true
+				break
+			}
+		}
+		return !isFilter
+	}
+	if len(ds.excludeExp) > 0 {
+		isExclude := false
+		for _, filter := range ds.includeExp {
+			if "" == strings.TrimSpace(filter) {
+				continue
+			}
+			if idx := strings.Index(name, allExp); idx > -1 {
+				filter = strings.ReplaceAll(filter, allExp, ".*")
+				if isMatch, er := regexp.MatchString(filter, name); er == nil && isMatch {
+					isExclude = true
+					break
+				}
+			} else if name == filter {
+				isExclude = true
+				break
+			}
+		}
+
+		ignore = isExclude
+	}
+	return ignore
+}
+
+func (ds *DirScanner) BaseDir() string {
+	return ds.baseDir
+}
+
 func NewDirScanner(vDir string) *DirScanner {
 	ds := &DirScanner{}
-	ds.baseDir = vDir
+	ds.baseDir = StdDir(vDir)
 	return ds
 }